Create base class for drupal contexts.
