+++ src/Behat/Context/DrupalExtendedContext.php
@@ -11,7 +11,6 @@ namespace Metadrop\Behat\Context;
 
 use Behat\Behat\Context\SnippetAcceptingContext;
 use Behat\Gherkin\Node\TableNode;
-use NuvoleWeb\Drupal\DrupalExtension\Context\RawDrupalContext;
 
 class DrupalExtendedContext extends RawDrupalContext implements SnippetAcceptingContext {
 
@@ -397,21 +396,6 @@ class DrupalExtendedContext extends RawDrupalContext implements SnippetAccepting
    public function userWithMailNotExists($mail) {
      $this->userWithMailExists($mail, FALSE);
    }
-
-  /**
-   * Overrides \Drupal\Driver\Cores\AbstractCore::expandEntityFields method.
-   *
-   * That method is protected and we can't use it from this context.
-   */
-  protected function expandEntityFields($entity_type, \stdClass $entity) {
-    $field_types = $this->getCore()->getEntityFieldTypes($entity_type);
-    foreach ($field_types as $field_name => $type) {
-      if (isset($entity->$field_name)) {
-        $entity->$field_name = $this->getCore()->getFieldHandler($entity, $entity_type, $field_name)
-          ->expand($entity->$field_name);
-      }
-    }
-  }
 
   /**
    * Checks that text appears before another text.
